widget.dart and fetchHistorical up to date, fixed loading error in historical graph
